crud entity

+++ src/main/java/com/example/SWP391/repository/FeedbackRepository.java
@@ -4,6 +4,6 @@ import com.example.SWP391.entity.Feedback;
 import com.example.SWP391.entity.License;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FeedbackRepository extends JpaRepository<Feedback, Integer> {
-    Feedback findFeedbackByFeedbackId(int feedbackId);
+public interface FeedbackRepository extends JpaRepository<Feedback, Long> {
+    Feedback findFeedbackByFeedbackId(long feedbackId);
 }
